refactor: extract routes from App.kt into com.splitit.routes namespace
